+++ src/Services/IdentityService/IdentityService.BLL/IdentityService.BLL.csproj
@@ -17,8 +17,4 @@
       <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="7.1.2" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="DTOs\Requests\" />
-    </ItemGroup>
-
 </Project>
