+++ src/Services/IdentityService/IdentityService.BLL/Providers/Implementation/TokenProvider.cs
@@ -2,13 +2,13 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Text;
+using IdentityService.BLL.Models.Results;
 using IdentityService.BLL.Providers.Interfaces;
-using IdentityService.BLL.Providers.Models;
 using IdentityService.Domain.Entities;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
-using TokenOptions = IdentityService.BLL.Providers.Options.TokenOptions;
+using TokenOptions = IdentityService.BLL.Models.Options.TokenOptions;
 
 namespace IdentityService.BLL.Providers.Implementation;
